app/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
wand
nltk
numpy
opencv-python-headless
gtts